requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
python-gnupg==0.5.1
qrcode[pil]==7.4.2
Pillow==10.0.0
Werkzeug==2.3.7

# SSH honeypot dependencies
paramiko==3.4.0
# google-generativeai==0.3.2
cryptography==41.0.7
python-dotenv==1.0.0
# tenacity==8.2.2